// lib/api/generated/models/PodcastResponseDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type PodcastResponseDto = {
    id: number;
    title: string;
    description: string;
    duration: string;
    audioUrl: string;
    imageUrl: string;
    views: number;
    createdAt: string;
    publisher: Record<string, any>;
};